Added a 'suggestions' page that the user is now redirected to after login instead of the 'dashboard' page.
The 'Suggestions' object retrievew twitter friends using an access token and secret from the configuration.json file, this should be
changed to use the token returned from passport after login. Nothing is
done with the retrieved data yet.

Created a function for logging the remaining rate limit and the time it
will be reset.

Renamed app.js back to index.js so it can be run automatically by
debugger.
Moved configuration-reader module to top level as it is now used by more
than one other module.