Split code into client and server subdirectories.
Used the create-react-app command to make use of the development features such as hot reloading that come with it.
It turns out these niceties come at a cost as there are a lot of dependencies and further complications when developing and testing a React app created with create-react-app and a server at the same time.
Some googling threw up a couple of articles that agreed on separating the client into its own app and then using a command line tool called 'concurrency' to bring both the development server for the client and the backend API server up at the same time and have the development server proxy requests to the backend to workaround CORS restrictions.
This seems insanely complex to me and I wonder how it has come to this but I do like the idea of developing and testing the client and server separately so some good may come of it.