Fix name and image edits not updating cache
The really weird thing about this was that profile edits were correctly
being returned from queries that hit the cache.
It makes sense that without an ID being returned for GetPerson
the cache was falling back to the last result for that query
rather than being able to hit the normalised cache.
What doesn't make sense is how it was able to use the updated
profiles that were only present in the mutation variables.